[FEATURE] Lower prio of command execution
To be able to fit in other scripts it is required
to make these commands execute at a later point
and at the same time unify their priorities.

The order of registration also results in the order of execution,
thus we can use the same prios for all commands here.